Pass the counts to the exception without casting them
clang-tidy's google-runtime-int rejects the `unsigned long long` and
`long long` casts on the two format arguments. They were never needed:
ExceptionFormatValue takes idx_t and int64_t as they are, and passing an
idx_t straight to a `%llu` is what the rest of src/main does.

+++ src/main/db_instance_cache.cpp
@@ -115,7 +115,7 @@ shared_ptr<DuckDB> DBInstanceCache::GetInstanceInternal(const string &database, 
 					live_instance.reset();
 					if (connections > 0) {
 						throw ConnectionException("Database \"%s\" is still in use: %llu connection(s) are open on it",
-						                          database, (unsigned long long)connections);
+						                          database, connections);
 					}
 					throw ConnectionException("Database \"%s\" is still in use: no connections are open, but "
 					                          "something still holds it, such as an unfinished query result",
@@ -124,8 +124,8 @@ shared_ptr<DuckDB> DBInstanceCache::GetInstanceInternal(const string &database, 
 				grace = now + std::chrono::milliseconds(IN_USE_GRACE_MILLIS);
 			}
 			if (now > deadline) {
-				throw ConnectionException("Database \"%s\" did not finish shutting down within %lld seconds", database,
-				                          (long long)SHUTDOWN_WAIT_SECONDS);
+				throw ConnectionException("Database \"%s\" did not finish shutting down within %d seconds", database,
+				                          SHUTDOWN_WAIT_SECONDS);
 			}
 		}
 		D_ASSERT(!cache_entry);
